ulord/requirements.txt
ecdsa==0.13
pbkdf2
requests
six>=1.9.0
slowaes==0.1a1
appdirs==1.4.3
protobuf==3.2.0
jsonschema==2.6.0
keyring==10.4.0
itsdangerous
passlib
Flask
Flask-RESTful
Flask-WTF
Flask-Script
Flask-SQLAlchemy
Flask-HTTPAuth
flask-marshmallow
marshmallow-sqlalchemy
psycopg2
simplejson
jsonrpclib-pelix